pyproject.toml
[project]
name = "astro-chart-extractor"
version = "0.1.0"
description = "Natal chart extraction and rendering app with FastAPI and Swiss Ephemeris"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pyswisseph>=2.10.3.2",
  "httpx>=0.27.0",
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["backend/tests"]
